# beat/docs/swagger.yaml
basePath: /api
definitions:
...
  presenters.MetadataErrorResponse:
    description: presenters.MetadataErrorResponse
    properties:
      data:
        example: ""
        type: string
      error:
        example: error message
        type: string
      status:
        example: false
        type: boolean
    type: object
  presenters.MetadataListResponse:
    description: presenters.MetadataListResponse
    properties:
      data: {}
      error:
        example: ""
        type: string
      status:
        example: true
        type: boolean
    type: object
  presenters.MetadataSuccessResponse:
    description: presenters.MetadataSuccessResponse
    properties:
      data: {}
      error:
        example: ""
        type: string
      status:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: Deals Beats, Beat_Listened, filtering
  title: Fiber Beat Service
  version: "1.0"
paths:
...
  /metadata/genres:
    get:
      description: Returns a list of all available music genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all genres
      tags:
      - Metadata
  /metadata/genres/popular:
    get:
      description: Returns a list of currently popular genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get trending genres
      tags:
      - Metadata
  /metadata/instruments:
    get:
      description: Returns a list of all available instruments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all instruments
      tags:
      - Metadata
  /metadata/keys:
    get:
      description: Returns a list of all available musical keys
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all keys
      tags:
      - Metadata
  /metadata/mfccs:
    get:
      description: Returns a list of all available MFCC data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all MFCCs
      tags:
      - Metadata
  /metadata/moods:
    get:
      description: Returns a list of all available mood categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all moods
      tags:
      - Metadata
  /metadata/tags:
    get:
      description: Returns a list of all available tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all tags
      tags:
      - Tags
  /metadata/tags/byName/{name}:
    get:
      description: Returns tag details for the specified name (ONLY 1 TAG WITH SPECIFIC
        NAME)
      parameters:
      - description: Tag name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get tag by name
      tags:
      - Tags
  /metadata/tags/byNameLike/{name}:
    get:
      description: Returns all of the tags for the specified name like%
      parameters:
      - description: Tag name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get MANY tags by name LIKE
      tags:
      - Tags
  /metadata/tags/in_trend:
    get:
      description: Returns a list of popular genres. Takes beats that were created
        this month (today minus 30 days), counts, how frequently were they used in
        beat_genres table.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get trending tags
      tags:
      - Tags
  /metadata/tags/random:
    get:
      description: Returns a list of randomly selected tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get random tags
      tags:
      - Tags
  /metadata/timestamps:
    get:
      description: Returns a list of all available timestamps
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.MetadataListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenters.MetadataErrorResponse'
      summary: Get all timestamps
      tags:
      - Timestamp
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"